refactor: OpenShift.php runner
