--- EMS/common-bundle/src/Helper/Runner/RunnerInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace EMS\CommonBundle\Helper\Runner;
-
-interface RunnerInterface
-{
-    /**
-     * @param string[] $command
-     */
-    public function start(array $command): string;
-}
